Move from configure to parameter
