+++ fcoin/api/endpoint/configure.go
@@ -2,7 +2,6 @@ package endpoint
 
 import (
 	"fcoin_go_client/fcoin/api"
-	"fmt"
 )
 
 type EndpointConfigure struct {
@@ -14,104 +13,6 @@ type EndpointConfigure struct {
 	Proxy     string
 	ApiKey    string
 	SecretKey string
-}
-
-type Endpoint struct {
-	Symbol     string
-	Level      string
-	Resolution string
-	Side       string
-	Price      string
-	Amount     string
-	Type       string
-	States     string
-	PageBefore string
-	PageAfter  string
-	PerPage    string
-	OrderId    string
-}
-
-type EndpointOption func(*Endpoint)
-type EndpointsOptions []EndpointOption
-
-func Symbol(s string) EndpointOption {
-	return func(e *Endpoint) {
-		e.Symbol = s
-	}
-}
-
-func Level(l string) EndpointOption {
-	return func(e *Endpoint) {
-		e.Level = l
-	}
-}
-
-func Resolution(r string) EndpointOption {
-	return func(e *Endpoint) {
-		e.Resolution = r
-	}
-}
-
-func Side(s string) EndpointOption {
-	return func(e *Endpoint) {
-		e.Side = s
-	}
-}
-
-func Price(p float32) EndpointOption {
-	return func(e *Endpoint) {
-		e.Price = fmt.Sprint(p)
-	}
-}
-
-func Amount(a float32) EndpointOption {
-	return func(e *Endpoint) {
-		e.Amount = fmt.Sprint(a)
-	}
-}
-
-func Type(t string) EndpointOption {
-	return func(e *Endpoint) {
-		e.Type = t
-	}
-}
-
-func States(s string) EndpointOption {
-	return func(e *Endpoint) {
-		e.States = s
-	}
-}
-
-func PageBefore(pb int) EndpointOption {
-	return func(e *Endpoint) {
-		e.PageBefore = fmt.Sprint(pb)
-	}
-}
-
-func PageAfter(pa int) EndpointOption {
-	return func(e *Endpoint) {
-		e.PageAfter = fmt.Sprint(pa)
-	}
-}
-
-func PerPage(pp int) EndpointOption {
-	return func(e *Endpoint) {
-		e.PerPage = fmt.Sprint(pp)
-	}
-}
-
-func OrderId(o string) EndpointOption {
-	return func(e *Endpoint) {
-		e.OrderId = o
-	}
-}
-
-func setEndpoint(opts EndpointsOptions) (ma *Endpoint) {
-	ma = &Endpoint{}
-	for _, opt := range opts {
-		opt(ma)
-	}
-	return
 }
 
 func apiConfig(c *EndpointConfigure) *(api.APIConfigure) {
